test(shared): split the source-state suite along its three seams
`shared/tests/test_source_state.py` reached 403 lines in #448 — a new crossing of
the 300-line guideline, which the bloat Stop-gate flagged after that PR had already
merged. Split rather than excepted, because the file had three genuinely separate
subjects and no reason to be one module:

* `test_source_state.py` (183) — the SHAPE contract: banner and JSON round-trips,
  run-id sanitisation, the anchored strip regex.
* `test_source_state_git.py` (131) — `resolve_git_state`: the half that is resolved
  by code, exercised against a real git repo including every unavailability path.
* `test_source_state_untrusted.py` (151) — the forgery regressions, each pinning a
  defect a review layer actually found (token collision, whitespace injection,
  commit validation, unsubstituted placeholder).

Behaviour-preserving: 45 test functions before, 45 after, verified by name-set
comparison against the merged file — none dropped, none added. All 81 cases pass.

Run-ID: iterate-2026-07-27-artifact-state-stamping
